http: move update customer status endpoint from admin server
