--- pkg/admin/docs/UpdateCustomerStatus.md
@@ -1,12 +1,0 @@
-# UpdateCustomerStatus
-
-## Properties
-
-Name | Type | Description | Notes
------------- | ------------- | ------------- | -------------
-**Comment** | **string** | Free form text for admin users | [optional] 
-**Status** | [**CustomerStatus**](CustomerStatus.md) |  | 
-
-[[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
-
-
