--- pkg/admin/model_update_customer_status.go
@@ -1,17 +1,0 @@
-/*
- * Customers Admin API
- *
- * Customers focuses on solving authentic identification of humans who are legally able to hold and transfer currency within the US. Primarily this project solves [Know Your Customer](https://en.wikipedia.org/wiki/Know_your_customer) (KYC), [Customer Identification Program](https://en.wikipedia.org/wiki/Customer_Identification_Program) (CIP), [Office of Foreign Asset Control](https://www.treasury.gov/about/organizational-structure/offices/Pages/Office-of-Foreign-Assets-Control.aspx) (OFAC) checks and verification workflows to comply with US federal law and ensure authentic transfers. Also, Customers has an objective to be a service for detailed due diligence on individuals and companies for Financial Institutions and services in a modernized and extensible way.
- *
- * API version: v1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package admin
-
-// UpdateCustomerStatus struct for UpdateCustomerStatus
-type UpdateCustomerStatus struct {
-	// Free form text for admin users
-	Comment string         `json:"comment,omitempty"`
-	Status  CustomerStatus `json:"status"`
-}
