+++ pkg/customers/approval_test.go
@@ -7,14 +7,13 @@ package customers
 import (
 	"bytes"
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 
+	"github.com/gorilla/mux"
 	"github.com/moov-io/base"
-	"github.com/moov-io/base/admin"
+	"github.com/stretchr/testify/require"
 
 	"github.com/moov-io/customers/pkg/client"
 
@@ -28,67 +27,32 @@ func TestCustomers__updateCustomerStatus(t *testing.T) {
 			Status:     "none",
 		},
 	}
-	searcher := createTestOFACSearcher(repo, nil)
-	ssnRepo := &testCustomerSSNRepository{}
+	router := mux.NewRouter()
+	AddCustomerRoutes(log.NewNopLogger(), router, repo, testCustomerSSNStorage(t), createTestOFACSearcher(repo, nil))
 
-	svc := admin.NewServer(":10001")
-	defer svc.Shutdown()
-	AddApprovalRoutes(log.NewNopLogger(), svc, repo, ssnRepo, searcher)
-	go svc.Listen()
+	updateStatusRequest := client.UpdateCustomerStatus{
+		Status:  "ReceiveOnly",
+		Comment: "test comment",
+	}
+	payload, err := json.Marshal(&updateStatusRequest)
+	require.NoError(t, err)
 
-	body := strings.NewReader(`{"status": "ReceiveOnly", "comment": "test comment"}`)
-	req, err := http.NewRequest("PUT", "http://localhost:10001/customers/foo/status", body)
+	req, err := http.NewRequest("PUT", "/customers/_id_/status", bytes.NewReader(payload))
 	if err != nil {
 		t.Fatal(err)
 	}
 	req.Header.Set("x-organization", "test")
 	req.Header.Set("x-request-id", "test")
 
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatal(err)
-	}
-	respBody, _ := ioutil.ReadAll(resp.Body)
-
-	if resp.StatusCode != http.StatusOK {
-		t.Errorf("bogus HTTP status: %d: %v", resp.StatusCode, string(respBody))
-	}
+	res := httptest.NewRecorder()
+	router.ServeHTTP(res, req)
+	require.Equal(t, http.StatusOK, res.Code)
 
 	var customer client.Customer
-	if err := json.NewDecoder(bytes.NewReader(respBody)).Decode(&customer); err != nil {
-		t.Fatal(err)
-	}
-	if customer.CustomerID == "" {
-		t.Errorf("missing customer JSON: %#v", customer)
-	}
-	if repo.updatedStatus != client.RECEIVE_ONLY {
-		t.Errorf("unexpected status: %s", repo.updatedStatus)
-	}
-}
-
-func TestCustomers__updateCustomerStatusFailure(t *testing.T) {
-	repo := &testCustomerRepository{}
-	searcher := createTestOFACSearcher(repo, nil)
-	ssnRepo := &testCustomerSSNRepository{}
-
-	w := httptest.NewRecorder()
-	req := httptest.NewRequest("GET", "/customers/foo/status", nil)
-	updateCustomerStatus(log.NewNopLogger(), repo, ssnRepo, searcher)(w, req)
-	w.Flush()
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("bogus HTTP status: %d", w.Code)
-	}
-
-	// try the proper HTTP verb
-	w = httptest.NewRecorder()
-	req = httptest.NewRequest("PUT", "/customers/foo/status", nil)
-	updateCustomerStatus(log.NewNopLogger(), repo, ssnRepo, searcher)(w, req)
-	w.Flush()
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("bogus HTTP status: %d", w.Code)
-	}
+	err = json.NewDecoder(res.Body).Decode(&customer)
+	require.NoError(t, err)
+	require.NotNil(t, customer)
+	require.Equal(t, updateStatusRequest.Status, repo.updatedStatus)
 }
 
 func TestCustomers__containsValidPrimaryAddress(t *testing.T) {
